* 'master' of github.com:scylladb/scylla: (58 commits)
  database: Use nop_large_partition_handler to avoid self-reporting
  cql3: alias single_column_primary_key_restrictions
  cql3: remove KeyType template from statement_restrictions
  cql3: remove template from primary_key_restrictions
  cql3: remove forwarding_primary_key_restrictions
  build: fix libdeflate object file corruption during parallel build
  cql3: really ensure retrieval of columns for filtering
  tools: toolchain: update docker build command to include --no-cache
  install-dependencies.sh: add packages that will be needed by scylla-python3
  build: Fix link order for DPDK
  tests: Split out `sstable_datafile_test`
  build: Remove unnecessary inclusion
  tests: Fix use-after-free errors in static vars
  build: Remove Seastar internals
  build: Only use Seastar flags from pkg-config
  build: Query Seastar flags using pkg-config
  hinted handoff: cache column family mappings for segments that were not sent out in full
  hinted handoff: add a "discarded" metric
  scylla_setup: make sure it works (again) in interactive mode
  tests: cql_query_test add test_time_overflow
  ...

+++ tests/sstable_test.cc
@@ -317,7 +317,7 @@ public:
     }
 
     virtual proceed consume_cell(bytes_view col_name, bytes_view value,
-            int64_t timestamp, int32_t ttl, int32_t expiration) override {
+            int64_t timestamp, int64_t ttl, int64_t expiration) override {
         BOOST_REQUIRE(ttl == 0);
         BOOST_REQUIRE(expiration == 0);
         switch (count_cell) {
@@ -433,7 +433,7 @@ public:
         return proceed::yes;
     }
     virtual proceed consume_cell(bytes_view col_name, bytes_view value,
-            int64_t timestamp, int32_t ttl, int32_t expiration) override {
+            int64_t timestamp, int64_t ttl, int64_t expiration) override {
         count_cell++;
         return proceed::yes;
     }
@@ -585,7 +585,7 @@ public:
     }
 
     virtual proceed consume_cell(bytes_view col_name, bytes_view value,
-            int64_t timestamp, int32_t ttl, int32_t expiration) override {
+            int64_t timestamp, int64_t ttl, int64_t expiration) override {
         switch (count_cell) {
         case 0:
             // The silly "cql row marker" cell
